Endpoint to update text in images

+++ src/routes/imageRoutes.js
@@ -23,9 +23,33 @@ router.get('/:id', checkAuth, async(req, res) => {
   res.json(images);
 });
 
+router.put('/:id', checkAuth, async(req, res) => {
+  const { id } = req.params;
+  const image = await Image.findById(id);
+  
+  if(!image) {
+      const error = new Error('Imagen no encontrada');
+      return res.status(404).json({ msg: error.message});
+  }
+
+  image.text = req.body.text;
+  try {
+      const savedImage = await image.save();
+      res.json({message: "Imagen guardada", _id: savedImage.id, text: savedImage.text })
+  } catch (error) {
+      return res.status(500).json({ msg: error.message + " - Contacte al administrador" });
+  }
+});
+
 router.delete('/:id', checkAuth, async(req, res) => {
   const { id } = req.params;
-  const image = Image.findById(id);
+  const image = await Image.findById(id);
+
+  if(!image) {
+    const error = new Error('Imagen no encontrada');
+    return res.status(404).json({ msg: error.message});
+  }
+
   try {
     await image.deleteOne();
     res.json({ msg: "Imagen Eliminada"});
